feat: add Django models for stories, story update logs, and story credits.
